added delete article protection

+++ app/controllers/articles_controller.rb
@@ -50,9 +50,17 @@ class ArticlesController < ApplicationController
   end
 
   def destroy
-    if @article.destroy
-      flash[:success] = "Article has been deleted"
+    unless @article.user == current_user
+      flash[:danger] = "You can only delete your own article"
       redirect_to articles_path
+    else
+      if @article.destroy
+        flash[:success] = "Article has been deleted"
+        redirect_to articles_path
+      else
+        flash.now[:danger] = "Article has not been updated"
+        render :edit
+      end
     end
   end
